% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidCluster-package.R
\docType{package}
\name{covidCluster-package}
\alias{covidCluster}
\alias{covidCluster-package}
\title{Automated Cluster Detection in Shelby County Health Department Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

covidCluster is designed to detect clusters of
    COVID-19 cases across space and time. It does so by analyzing data on
    the whereabouts of interviewed cases and identifying links between
    those cases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jesse-smith.github.io/covidCluster/}
  \item \url{https://github.com/jesse-smith/covidCluster}
}

}
\author{
\strong{Maintainer}: Jesse Smith \email{jesse.smith@shelbycountytn.gov} (\href{https://orcid.org/0000-0001-7375-2463}{ORCID})

}
\keyword{internal}
