% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{create_table}
\alias{create_table}
\title{Create a One-Way Table from Multiple Variables}
\usage{
create_table(.data, ..., to = NULL, infreq = NULL, show_missing_levels = FALSE)
}
\description{
\code{pivot_table} works similarly to \code{link{create_table}}, but is
designed to "pivot" the input into long format before summarizing. This is
most useful when you'd like to treat multiple variables as a single variable
in the summary table.
}
