% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-coviData.R
\name{download_integrated_data}
\alias{download_integrated_data}
\title{Wrappers Around coviData Function}
\usage{
download_integrated_data(
  date = Sys.Date(),
  api_token = Sys.getenv("redcap_IDT_token"),
 
    directory = "V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/Integrated data tool Case Interviews/",
  new_file = paste0("integrated_data_", date, ".csv"),
  convert = FALSE,
  force = FALSE
)
}
\arguments{
\item{date}{The date with which to name \code{new_file}}

\item{api_token}{The API access token for the Integrated Data Tool REDcap
project. This should be stored in a \emph{.Renviron} file; see
\link[coviData]{using_Renviron} for details.}

\item{directory}{A string specifying the directory in which to save the data;
this should usually be left alone}

\item{new_file}{A string specifying the file name to save the data under;
this should always end in \emph{.csv} and should usually be left alone}

\item{convert}{A logical indicating whether to open in Excel for conversion
to \emph{.xlsx} format (or another file format)}

\item{force}{A logical indicating whether to overwrite an existing file with
the same date; the default is \code{FALSE} and should usually be left alone}
}
\description{
\code{coviData} is the package for down/loading, munging, and performing
basic analysis on COVID-19 data for Shelby County. It includes functionality
for downloading data from the integrated data tool, which is re-exported in
covidCluster.
}
