% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fct_cut}
\alias{fct_cut}
\title{"Cut" a Numeric Variable into Categories}
\usage{
fct_cut(.x, breaks = quantile(.x), open = FALSE, ordered = TRUE)
}
\arguments{
\item{.x}{A numeric vector, or a object that can be coerced to one. If
\code{.x} is a factor, the level \emph{labels} will be coerced, not the
underlying integer representation.}

\item{breaks}{A numeric vector of break points for the categories of
\code{.x}}

\item{open}{Should the ends of \code{breaks} be treated as open or closed? In
other words, should there be a category for values in \code{.x} below the
lowest value of \code{breaks} (\code{.x < min(breaks)}) or above the
highest value of \code{breaks} (\code{.x >= max(breaks)})? \code{TRUE}
means both are open; \code{FALSE} means both are closed (the default).
Behavior for each end can be controlled by passing a logical vector of
length 2 where the first value controls the lower end and the second
controls the upper end (e.g. \code{open = c(FALSE, TRUE)} for count data).}

\item{ordered}{Should the resulting factor be ordered? Since numeric data is
inherently ordered, the default is \code{TRUE}.}
}
\value{
A \code{factor}
}
\description{
\code{fct_cut} converts numeric variables to a factor using the \code{breaks}
provided. It is built on \code{\link[base]{cut}}, but does some additional
coercion before passing to \code{cut} and labels the output in more
lay-friendly manner. Note that intervals in \code{fct_cut} are closed on the
left (they include the lower value and exclude the higher - e.g.
\code{1 <= .x < 2}); this is the opposite of \code{cut}'s default behavior.
}
