% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-coviData.R
\name{load_integrated_data}
\alias{load_integrated_data}
\title{Load Data from the Integrated Data Tool REDcap Project}
\usage{
load_integrated_data(
  date = Sys.Date(),
  directory = paste0("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "Integrated data tool Case Interviews/data/"),
  ext = "csv"
)
}
\arguments{
\item{date}{A \code{Date} object indicating the date of the data file to load}

\item{directory}{A string specifying the directory in which to search for
files containing \code{date}}

\item{ext}{The file type to search for}
}
\value{
A \code{link[tibble]{tibble}} containing the data from the specified
file
}
\description{
\code{load_integrated_data} loads data from the Integrated Data Tool REDcap
project. It is designed to be used in conjuction with
\link[coviData]{download_integrated_data}.
}
