% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_bar}
\alias{plot_bar}
\title{Create Bar Chart from Table}
\usage{
plot_bar(
  .table,
  .by = NULL,
  incl_missing = FALSE,
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  legend = FALSE,
  legend_title = title,
  bottom_legend = TRUE,
  palette = "blue",
  background = "gray93"
)
}
\arguments{
\item{.table}{The output of \code{\link{create_table}} which is a one-way
\code{\link[janitor]{tabyl}} encoded to assist visualization}

\item{.by}{The variable to chart; defaults to the first column of
\code{.table}, which will always be correct if the output is from a
\code{*_table} function}

\item{incl_missing}{Should the chart include missing values as a category?}

\item{title}{The title of the chart; defaults to a title-case version of the
variable name}

\item{x_lab}{Label for the x-axis; defaults to no label}

\item{y_lab}{Label for the y-axis; defaults to no label}

\item{legend}{Should a legend be displayed? Bar charts should usually not
require a legend, so the default is \code{FALSE}}

\item{legend_title}{The title of the chart legend; defaults to the chart
title}

\item{bottom_legend}{Should the legend be positioned at the bottom or on the
right side? Default is bottom.}

\item{palette}{The color palette to use; this is a material design color
palette generated by \code{\link[ggsci]{pal_material}}; see that
function for details and available choices}

\item{background}{What should the background color be? Defaults to a light
gray to lessen eye strain.}
}
\value{
A ggplot object
}
\description{
\code{create_bar} takes output from one of the \code{*_table} functions and
produces a bar chart.
}
