% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-coviData.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{download_integrated_data}
\alias{load_integrated_data}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{coviData}{\code{\link[coviData]{download_integrated_data}}, \code{\link[coviData]{load_integrated_data}}}
}}

