% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-coviData.R
\name{load_integrated_data}
\alias{load_integrated_data}
\title{Wrappers Around coviData Function}
\usage{
load_integrated_data(
  date = Sys.Date(),
  directory = paste0("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "Integrated data tool Case Interviews/data/"),
  ext = "csv"
)
}
\arguments{
\item{date}{A \code{Date} object indicating the date of the data file to load}

\item{directory}{A string specifying the directory in which to search for
files containing \code{date}}

\item{ext}{The file type to search for}
}
\description{
\code{coviData} is the package for down/loading, munging, and performing
basic analysis on COVID-19 data for Shelby County. It includes functionality
for loading data from the integrated data tool, which is re-exported in
covidCluster.
}
