% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pie}
\alias{plot_pie}
\title{Create Pie Chart from Table}
\usage{
plot_pie(
  .table,
  .by = NULL,
  incl_missing = FALSE,
  title = NULL,
  legend_title = title,
  palette = "blue",
  bottom_legend = TRUE,
  background = "gray93",
  dodge = 0
)
}
\arguments{
\item{.table}{The output of \code{\link{create_table}}, which is a one-way
\code{\link[janitor]{tabyl}} encoded to assist visualization}

\item{.by}{The variable to chart; defaults to the first column of
\code{.table}}

\item{incl_missing}{Should the chart include missing values as a category?}

\item{title}{The title of the chart; defaults to a title-case version of the
variable name}

\item{legend_title}{The title of the chart legend; defaults to the chart
title}

\item{palette}{The color palette to use; this is a material design color
palette generated by \code{\link[ggsci]{pal_material}}; see that
function for details and available choices}

\item{bottom_legend}{Should the legend be positioned at the bottom or on the
right side?}

\item{background}{What should the background color be? Defaults to a light gray.}

\item{dodge}{Change the position of the labels along the radius of the
circle; useful for separating overlapping labels. A value of \code{0}
leaves the labels at the mid-point of the pie chart radius; a value of
\code{1} spaces them evenly across the radius. Values in-between \code{0}
and \code{1} spread the labels out across that fraction of the radius
(centered on the midpoint).}
}
\value{
A ggplot object
}
\description{
\code{create_pie} takes output from one of the \code{covidCluster::*_table}
functions and produces a pie chart.
}
